move key sending on linux to own daemon executable
on linux you usually need root access to emit key events. to avoid the
need to request root rights for the whole server application with the gui,
the key sender is now moved to its own lightweight executable. This
executable can be executed with root rights and will open a network server
that listens for events on localhost. The gui will then send the requests
for key pressing to this server.

Fixes #16 and #17